backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
python-multipart
numpy
pillow
scikit-learn